Add module 3.5

+++ Basic Data Structures/Module 3.5/Summation_from_1_to_N.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 int main()
 {
-    int n;
+    long long int n;
     cin >> n;
     cout << (n * (n + 1)) / 2;
     return 0;
